# docs/api/glmfitlins_openapi.yaml
openapi: 3.0.3
info:
  title: GLM FitLins API
  version: 0.1.0
  description: Draft OpenAPI spec for the BR-KG GLM FitLins endpoints.
servers:
  - url: /api/glmfitlins
paths:
...
  /search:
    get:
      summary: Search datasets, contrasts, and constructs
      tags: [Search]
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: type
          schema:
            type: string
            enum: [dataset, contrast, construct]
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Missing query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /stats:
    get:
      summary: Get GLM FitLins statistics
      tags: [Stats]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    DatasetSummary:
      type: object
      properties:
        id: { type: string }
        dataset_id: { type: string }
        name: { type: string }
        doi: { type: string }
        source: { type: string }
        contrast_count: { type: integer }
...
    ContrastSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        task_label: { type: string }
        dataset_id: { type: string }
        construct_count: { type: integer }
...
    ConstructSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        node_id: { type: string }
        usage_count: { type: integer }
        avg_confidence: { type: number }
...
    SearchResponse:
      type: object
      properties:
        datasets:
          type: array
          items:
            $ref: "#/components/schemas/DatasetSummary"
        contrasts:
          type: array
          items:
            $ref: "#/components/schemas/ContrastSummary"
        constructs:
          type: array
          items:
            $ref: "#/components/schemas/ConstructSummary"
        total:
          type: object
          properties:
            datasets: { type: integer }
            contrasts: { type: integer }
            constructs: { type: integer }
            all: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
...
    StatsResponse:
      type: object
      properties:
        database:
          type: object
          properties:
            total_nodes: { type: integer }
            total_relationships: { type: integer }
        glmfitlins:
          type: object
          properties:
            datasets: { type: integer }
            contrasts: { type: integer }
            constructs: { type: integer }
            annotations: { type: integer }
            subject_groups: { type: integer }
            subjects: { type: integer }
            phenotypes: { type: integer }
        relationships:
          type: object
          properties:
            involves_construct: { type: integer }
            includes: { type: integer }
            has_subject: { type: integer }
            has_phenotype: { type: integer }
        confidence_stats:
          type: object
...